requirements.txt
altgraph==0.17
auto-py-to-exe==2.10.0
biopython==1.79
bottle==0.12.19
bottle-websocket==0.2.9
cffi==1.14.6
Eel==0.12.4
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.1
numpy==1.21.1
pefile==2021.5.24
pycparser==2.20
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
pywin32-ctypes==0.2.0
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
